requirement.txt
asgiref==3.5.0
beautifulsoup4==4.11.1
Django==3.0.3
django-bootstrap4==22.1
djangorestframework==3.13.1
factory-boy==3.2.1
Faker==13.12.0
psycopg2==2.8.6
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
